requirements.txt
Flask
flask-mysql-connector
python-dotenv
flask-session
flask-bcrypt